webhook-channel/pyproject.toml
[project]
name = "webhook-channel"
version = "0.0.2"
dependencies = [
    "mcp==1.26.0",
    "uvicorn==0.42.0",
    "starlette==0.52.1",
    "slack_sdk==3.40.1",
    "aiohttp==3.13.3",
    "python-dotenv==1.2.2",
]

[project.scripts]
webhook = "webhook:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["webhook.py", "channels/"]
